% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrpi.R
\docType{package}
\name{ccrpi}
\alias{ccrpi}
\alias{ccrpi-package}
\title{Calculate Georgia (U.S.) Schools' College & Career Ready Performanc Index (CCRPI) Scores in R}
\description{
R Package for computing CCRPI scores for GA schools (based on the [2018 Redesigned CCRPI Scoring Model](http://www.gadoe.org/Curriculum-Instruction-and-Assessment/Accountability/Pages/For-Educators.aspx))
}
