% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcmbl.R
\docType{data}
\name{lgrps}
\alias{lgrps}
\title{---
title: "Fulton County Schools CCRPI 2017 Replicated using 2018 Scoring Model: Closing Gaps Baseline Data"
author: "Riley Smith-Hunter"
date: "Last Updated: `r format(Sys.Date(), '%d %b %Y')`"
---}
\format{An object of class \code{list} of length 10.}
\usage{
lgrps
}
\description{
-----
}
\details{
`r tufte::newthought("Note:")` This file only needs to be run once to generate the subgroups' baseline scores datafile ('_**`data/CM-bySubgroup-2016-new.csv`**_'). _This script was originally written and implemented on `2018-06-08`_;

------

# General Setup
}
\keyword{datasets}
