% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcgpts.R
\name{Rcgpts}
\alias{Rcgpts}
\title{-----}
\usage{
Rcgpts(x, df.baseline, df.target, subject,
  in_vivo_qc = getOption("in_vivo_qc"))
}
\description{
# `Rcgpts()`: Algorithm for Computing Closing Gaps Points Earned by Subgroup
}
