% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcgpts_post.R
\name{Rcgpts_post}
\alias{Rcgpts_post}
\title{# Post-Process results of '`Rcgpts()`' Output Data}
\usage{
Rcgpts_post(x)
}
\arguments{
\item{x}{todo}
}
\description{
# Post-Process results of '`Rcgpts()`' Output Data
}
