% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcgpts_sgt.R
\name{Rcgpts_sgt}
\alias{Rcgpts_sgt}
\title{Pre-Process Input Data for `Rcgpts()` (defined later)}
\usage{
Rcgpts_sgt(xw, category_var = "reporting.category",
  baseline_vars_pattern = "baseline\\\\.", group_labs = c("all", "ai", "as",
  "bl", "hp", "mr", "wh", "ed", "el", "swd"))
}
\arguments{
\item{...}{Additional arguments \emph{not yet implemented}.}
}
\description{
Pre-Process Input Data for `Rcgpts()` (defined later)
}
