% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcmbl.R
\name{Rcmbl}
\alias{Rcmbl}
\title{Generate Subgroups' \emph{Content Mastery} Baseline Data (todo)}
\usage{
Rcmbl(bl.subj, gradecluster, subject, ...)
}
\description{
Generate Subgroups' \emph{Content Mastery} Baseline Data (todo)
}
