% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcgpts.R
\docType{data}
\name{rec.wgt}
\alias{rec.wgt}
\title{## `Rrecwgt()` & `Rnflags()`}
\format{An object of class \code{character} of length 1.}
\usage{
rec.wgt
}
\description{
## `Rrecwgt()` & `Rnflags()`
}
\keyword{datasets}
