% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relp.R
\name{Relp}
\alias{Relp}
\title{Compute \emph{Progress} Component EL Indicator Points (todo)}
\usage{
Relp(x, ids = c(school = "school.id", student = "gtid", group =
  "student.grade.level"))
}
\description{
Compute \emph{Progress} Component EL Indicator Points (todo)
}
