% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rindicators.R
\name{Rindicators}
\alias{Rindicators}
\title{Compute benchmarked indicator points (todo)}
\usage{
Rindicators(x, idvar, v.names, timevar, benchmark = 1)
}
\description{
Compute benchmarked indicator points (todo)
}
