% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlxmid.R
\name{Rlxmid}
\alias{Rlxmid}
\title{Compute Lexile Midpoints & \emph{Readiness} Component Lexile Indicator Performance Points}
\usage{
Rlxmid(x, grades, gradevar = "student.grade.level", midpoints,
  fayfilter = TRUE)
}
\description{
Compute Lexile Midpoints & \emph{Readiness} Component Lexile Indicator Performance Points
}
