% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprgpts.R
\name{Rprgpts}
\alias{Rprgpts}
\title{Compute Student Growth Points (SGPs) for one subject area and across all grade levels at once (todo).}
\usage{
Rprgpts(x, subjectarea, fayfilter = TRUE, assessmenttype = c("EOG", "EOC",
  "GAA"), ids = c("school.id", "gtid"), ptsvar = "x2018.points")
}
\description{
Compute Student Growth Points (SGPs) for one subject area and across all grade levels at once (todo).
}
