% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcgpts.R, R/Rcmbl.R
\docType{data}
\name{lgrps}
\alias{lgrps}
\alias{lgrps}
\title{-----}
\format{An object of class \code{list} of length 10.}
\usage{
lgrps

lgrps
}
\description{
# `Rcgpts()` Universal Setup

-----
}
\details{
`r tufte::newthought("Note:")` This file only needs to be run once to generate the subgroups' baseline scores datafile ('_**`data/CM-bySubgroup-2016-new.csv`**_'). _This script was originally written and implemented on `2018-06-08`_;

------

# General Setup
}
\keyword{datasets}
