% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rach.R
\name{Rach}
\alias{Rach}
\title{\emph{Content Mastery} Component Score Calculation.}
\usage{
Rach(x, gradeband, grade_var = "student.grade.level", subject, subject_var,
  subjectcodes, assessment_type_var, assessment_type_codes = c("EOG", "EOC",
  "GAA"), fay_var, fay_code = "Y",
  performance_code_var = "osa.performance.code",
  valid_performance_codes = c("BEG", "DNM", "DEV", "PRO", "ADV", "DIS"),
  rec.cmlvls = c("'BEG' = 'L1'; 'DNM' = 'L1'; 'DEV' = 'L2'; 'PRO' = 'L3'; 'ADV' = 'L3'; 'DIS' = 'L4'"),
  rec.cmpnts = c("'BEG' = 0; 'DNM' = 0; 'DEV' = 0.5; 'PRO' = 1; 'ADV' = 1; 'DIS' = 1.5"),
  group_var = NULL, group = NULL, return_new_x = TRUE, ...)
}
\arguments{
\item{x}{A \code{dataframe} or object coercible to a \code{dataframe} containing \emph{GA Milestones} data to be used in computing subject-level \emph{Content Mastery} indicator points.}

\item{gradeband}{A character vector of length 1 specifying the focal gradeband as one of the following: "ES" for elementary schools (grades 3 - 5), "MS" for middle schools (grades 6 - 8), or "HS" (grades 9 - 12).}

\item{grade_var}{A character vector of length 1 specifying the column name in \code{x} containing students' grade level}

\item{subject}{A character vector of length 1 specifying the focal subject area s one of the following: "ELA" for English/Language Arts, "MATH" for Mathematics, "SCI" for Science, or "SS" for Social Studies.}

\item{subject_var}{A character vector of length 1 specifying the column name in \code{x} containing the appropriate subject area codes.}

\item{subjectcodes}{A character vector specifying the actual codes used in \code{x$subject_var} for the subject area specified in \code{subject}.}

\item{assessment_type_var}{A character vector of length 1 specifying the column name in \code{x} containing the assessment types (e.g., "EOG", "EOC", "GAA", etc.).}

\item{assessment_type_codes}{A character vector specifying the acceptable values of \code{assessment_type_var} for use in computing content mastery scores.}

\item{fay_var}{A character vector of length 1 specifying the column name in \code{x} containing the \emph{FAY Participant} filter.}

\item{fay_code}{A character vector specifying acceptable values of \code{fay_var} for the \emph{FAY} filter.}

\item{performance_code_var}{A character vector of length 1 specifying the column name in \code{x} containing the assessment performance codes (e.g., "BEG", "DEV", "PRO", "DIS").}

\item{valid_performance_codes}{A character vector specifying the acceptable performance code values under \code{performance_code_var}.}

\item{rec.cmlvls}{A character vector of (pseudo) length 1 providing the recodes to be passed to \code{car::recode()} for the \emph{Content Mastery} achievement levels (see \emph{Details}).}

\item{rec.cmpnts}{A character vector of (pseudo) length 1 providing the recodes to be passed to \code{car::recode()} for the point values corresponding to the \emph{Content Mastery} achievement levels (see \emph{Details}).}

\item{group_var}{An \emph{optional} character vector specifying the column name in \code{x} containing the appropriate variable for the focal subgroup (see \emph{Details} and \code{\link[ccrpi]{Rcgpts}}).}

\item{group}{A character vector specifying the appropriate subgroup under \code{group_var} (ignored if \code{group_var} is \code{NULL}).}

\item{return_new_x}{logical. If \code{TRUE} (the default), the final returned value is a list containing (1) the manipulated version of the input dataframe (`x`), for QC purposes, and (2) the computed subject-level \emph{Content Mastery} points dataframe.}

\item{...}{Additional arguments \emph{not yet implemented}.}
}
\description{
Dynamically compute subject-area indicator points for the \emph{Content Mastery CCRPI} component, overall (per individual school, or across multiple schools within the same gradeband) or by subgroup(s).
}
