% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcgpts.R
\name{Rcgpts_sgt}
\alias{Rcgpts_sgt}
\title{------}
\usage{
Rcgpts_sgt(xw)
}
\description{
# '`cmbl`': Subgroups' Baseline Data
}
\details{
-----

# `Rcgpts_sgt()`: Pre-Process Input Data for `Rcgpts()` (defined later)
}
